# internal/model/subscription.go
package model

type Subscription struct {
	ID          int     `json:"id" db:"id"`
	ServiceName string  `json:"service_name" db:"service_name"`
	Price       int     `json:"price" db:"price"`
	UserID      string  `json:"user_id" db:"user_id"`
	StartDate   string  `json:"start_date" db:"start_date"`
	EndDate     *string `json:"end_date,omitempty" db:"end_date"`
}